PXC-2344 : PXC 5.7.24 hotfix (with PS-4716 hotfix)
Branch of PXC 5.7.24 with PS-4716 hotfix

(From PS-4716)
A quick and dirty solution to this deadlock would be to refactor PolyLock_lock_log class
so that it would only try to acquire mysql_bin_log.get_log_lock() for write operation.

A fix is available upstraem in 5.7.25, but that has not been released yet.